Rethink image upload storage collisions. Don't ever plan on raising an error from upload_to, unless Django's storage framework itself runs into a problem. But if we have to delegate to the storage framework to prevent collisions, it probably means the filename generation algorithm is inadequate, so mail a notice to the admins.
Before this, the collision test was causing a very weird error on an upload metadata test when running the whole test suite. The same images were getting reported as one set of IDs in one place, and a different set of IDs in another place. refresh_from_db() didn't seem to help. The problem went away when the collision test didn't throw an error from upload_to.