IMPORTANT FOR MIGRATIONS: READ THIS WHOLE COMMIT MESSAGE. Squash migrations in accounts, annotations, images, labels, and vision_backend.
This commit squashes migrations for the above apps. The next commit will delete the original migrations. To get up to speed, do the following:

1. Update your local branch to JUST BEFORE this commit. Run any migrations as necessary until `manage.py showmigrations` shows that you're caught up with migrations. If you get an "attribute name must be string" error, try running them in a different order (like vision_backend before annotations).

2. Update your local branch to THIS commit. Run `manage.py migrate` so that Django reads your migration history and registers the squashed migrations as run (since all the replaced migrations were run previously). Now when you do `manage.py runserver`, it shouldn't say you have unapplied migrations.

3. You're now free to update your local branch past this commit.

Explanation: Unfortunately, after creating the squashed migrations, many of the original migrations had to have their dependencies changed nonsensically in this commit. Otherwise Django just kept claiming there was a circular dependency between migrations. This is probably because the apps had circular dependencies between each other's models at one point or another in the migration history.

So after squashing, it was impossible to keep the original migrations as safe to run. The reason the original broken migrations are kept for this commit, is that it's necessary to let the original and squashed migrations co-exist for at least one commit. Only at that point, it's possible for `manage.py migrate` to register the squashed migrations as run, based on the original migrations' run history.