Replace celery with huey, another package for async tasks.
Previously tried to update celery to 5.1.2, but couldn't solve "ValueError: not enough values to unpack" when a periodic task tried to run.

While we're at it: 1) All huey tasks are now also CoralNet jobs, as it makes things a bit cleaner. 2) Removed the last of the 10-second task delays; queue_job() is a better way to guarantee that the tasks don't start till the view finishes.