Log classify-image events, so we can more reliably detect whether the current classifier has done classifications.
The most important effect is to fix a corner case where image-classifications are queued for the same images infinitely. This happened when the current and previous classifier predicted the same labels for ALL unconfirmed points in a source.

We do this logging with a new model in a new 'events' app. This has basically the same ambitions as 'newsfeed', but newsfeed hasn't been used at all yet, and now is as good a time as any to be pedantic about its details; hence the new app and model. If events ends up working out well, then newsfeed could be removed later.

There is some overlap between events and jobs, but for now they seem different enough to track separately. And old jobs aren't kept around for more than 30 days.